requirements.txt
Flask>=2.2
Pillow>=9.5
gunicorn>=20.0
